cloud-run/export-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
boto3==1.34.0
pydantic==2.5.3
python-multipart==0.0.6
google-cloud-firestore>=2.16.0
